[Resolve #40]separate Web and Console from QC
