+++ main.go
@@ -1,18 +1,13 @@
 package main
 
 import (
-	"github.com/Qitmeer/qitmeer-wallet/console"
+	"github.com/Qitmeer/qitmeer-wallet/commands"
 	"github.com/Qitmeer/qitmeer/log"
 	"os"
 )
-var rootCmd =console.RootCmd
 
-
-func init()  {
-	console.BindFlags()
-}
 func main() {
-	if err := rootCmd.Execute(); err != nil {
+	if err := commands.RootCmd.Execute(); err != nil {
 		log.Error(err.Error())
 		os.Exit(1)
 	}
